PQ:Refactor message queue and row exchange and process early detach workers.
PQ:Refactor message queue and row exchange and process early detach workers.
- move enum definition into corresponding class.
- don't need 2 events for message queue sender event and receiver event does not
  be used at same time.
- when a worker reports an error, kill other workers.
- change type of m_detached to std::atomic
- Remove a redundant call of NotifyPeer()
- Remove rows sent counter, it should not be in message queue.

Also fix an ASSERTION when a worker reports an INTERRUPTED in mtr case

When a worker reports an INTERRUPTED, leader also send a KILL to it which leads
leader ignore the INTERRUPTED. To resolve this, introduces a new worker state:
'Cleaning'. If a worker in this state, it skips termination from leader.