Implements a raw MDL group lock
Implements a raw MDL group lock

All threads of a parallel query will join same MDL group lock. If a thread holds
a lock, this lock will be granted to all threads of the group. We just checked
it in MDL_lock::can_grant_lock().

This implementation is not efficient. The checking same lock group could be a
bottleneck if there are so many sessions. This should be improved in future.

One possible solution:
- Add a list to maintain members of a lock group, then just check the list.
- Add a ticket list in group leader to maintain all fast path locks to enable
  fast path locks again.