Fix PQ: EXPLAIN ANALYZE misses nodes under MATERIALIZE on workers
Fix PQ: EXPLAIN ANALYZE misses nodes under MATERIALIZE on workers

The root cause is that the parameter 'timing_data' is not passed to
ExplainMaterializeAccessPath(). it leads to data of nodes under MATERALIZE is
missing but leader read serialized timing data when running into that node, then
garbage data is read.

Adds an asseration to avoid the situation occurs again and add a new test case
to test EXPLAIN and EXPLAIN ANALYZE for parallel query.