PQ: introduce an event service for fd based row channel or other communications
PQ: introduce an event service for fd based row channel or other communications

In current row channel implementation, we use an Event in RowExchange, we a
channel needs to wait it will wait on this event. If there is an income row, it
will be weak up by others, e.g. sender channel will weak up it in a shared memory
row channel.

For TCP or other FD based row channel, this commit introduce an event service
which poll on fds then send event to parallel query threads if there is a read /
write event.

Note, There are several known issues in current implementation, see sources for
details.