PQ:Don't change Item::marker in receate_materialized_table()
PQ:Don't change Item::marker in receate_materialized_table()

create_tmp_table() could create different length of buffer if item markers are
different. We use buffer of tmp table to send tuples to leader. And if we
changed markers of item when clone plan, receive buffer on leader and send
buffer on workers may be inconsistent.