PQ: introduce variable "parallel_query_switch" to turn parallel query capabilities on and/or off
PQ: introduce variable "parallel_query_switch" to turn parallel query capabilities on and/or off

As optimizer_switch to control optimizers, this variable includes switches for
parallel query. we can use it to turn certain parallel query capabilities on
and/or off.

Currently, there is only one flag (join) available.