Move parallel plan choosing and generating to Query_block::optimize()
Move parallel plan choosing and generating to Query_block::optimize()

This makes current query block's parallel plan is created after all subqueries
parallel plan generating. So we can determines whether a subquery should be
pre-evaluated. For example, If a subquery hasn't a parallel plan, we can just
push it down to workers. currently, we don't support that because there is no
subqueries support on workers.