PQ: Fix an assertion when generating a parallel materialize access path
PQ: Fix an assertion when generating a parallel materialize access path

It asserts that QEP_TAB::needs_duplicate_removal is in rewrite_materialize().
But it is not true because sorting could create materialize access path. see the
positions where the variable need_tmp_before_win is set. A mtr test case is
added for this issue.