PQ: Fix rebind field fail if a unpushed item refer to a duplicated field in select fields
PQ: Fix rebind field fail if a unpushed item refer to a duplicated field in select fields

We don't push down an item if it is a const item. It could failure to resolve
Item_field if a const item refer to some table fields that are already in query
block fields because in GenPartialFields() we omit the field if it is already in
pushed fields.

To fixed that, we also find target field by Field address after by Item ID.
For same reason, also remove an const_item() assert in resolve_view_ref() since
a Item_view_ref could be refered by a const item, but itself is not a const item.

This issue was found by sqlsmith.