Decorrelate the subquery by using window function.
reference: https://www.researchgate.net/publication/221214692_WinMagic_Subquery_Elimination_Using_Window_Aggregation

[feature] Decorrelate the subquery by using window function.

We go a further step to simply decorrelate the subquery. By making
use of extended window aggregation capabilities, we eliminate
redundant access to common tables referenced in the outer query
block and the subquery.

This technique can mainly be exploited for correlated subqueries.
It is possible to get a huge boost in performance for queries that
can exploit this technique, which we call winmagic. The algorithm
will improve the performance of query No.2 in TCPH by 5 times, and
improve the performance of query No.17 by 50%. No.11 and No.15 will
be supported later.

Add subquery_to_window=off to optimizer_switch and open it if needed.