fix ANALYZE temporary table when collect basic column statistics coredump
- Cause of the problem:
The temporary table cannot be opened twice in one query
Currently,ANALYZE operation use SELECT Query to get BASIC_COLUMN_STATISTICS statistics(min-max) info.
If the operation is ANALYZE temporary table,the temporary table will be open twice during ANALYZE when we get basic column stats, at this time there will be an error.
So we need to set the flag to allowed temporary table can be reopen before the Query execute.
- Solution:
If we need to get BASIC_COLUMN_STATISTICS statistics when performing an ANALYZE operation, we need to add the following operation:
1.check whether the table is a temporary table
2.if it is a temporary table,call mark_tmp_table_reuse to allow the temporary table open twice during Query