PQ: Fix EXPLAIN ANALYZE crash without execution of parallel subqueries.
PQ: Fix EXPLAIN ANALYZE crash without execution of parallel subqueries.

In original mysql, It does not create iterators for access pathes if one
subquery is not executed in a plan. But in parallel query, we creates fake
timing iterators for a parallel plan at end of parallel plan generating if we
run an EXPLAIN ANALYZE command. This leads to a crash if a parallel subquery is
not executed at all. To fix this, we move creation of fake timing iterator after
collector iterator is created. Also add check validation of timing data when
partial plan timing data reading.